+++ src/Selenium2Library/locators/windowmanager.py
@@ -24,7 +24,15 @@ class WindowManager(object):
 
     def select(self, browser, locator):
         assert browser is not None
-
+        if locator is not None:
+            if isinstance(locator, list):
+                self._select_by_excludes(browser, locator)
+                return
+            if locator.lower() == "self" or locator.lower() == "current":
+                return
+            if locator.lower() == "new" or locator.lower() == "popup":
+                self._select_by_last_index(browser)
+                return
         (prefix, criteria) = self._parse_locator(locator)
         strategy = self._strategies.get(prefix)
         if strategy is None:
@@ -53,21 +61,35 @@ class WindowManager(object):
 
     def _select_by_default(self, browser, criteria):
         if criteria is None or len(criteria) == 0 or criteria.lower() == "null":
-            browser.switch_to_window('')
+            browser.switch_to_window(browser.get_window_handles()[0])
             return
-
+        for handle in browser.get_window_handles():
+            browser.switch_to_window(handle)
+            if criteria == handle:
+                return
+            for item in browser.get_current_window_info()[2:4]:
+                if item.strip().lower() == criteria.lower():
+                    return
+        raise ValueError("Unable to locate window with handle or name or title or URL '" + criteria + "'")
+    
+    def _select_by_last_index(self, browser):
+        handles = browser.get_window_handles()
         try:
-            self._select_by_name(browser, criteria)
-            return
-        except ValueError: pass
+            if handles[-1] == browser.get_current_window_handle():
+                raise AssertionError("Unable to get new window from last index. Please use '@{ex}= | List Windows' + new window trigger + 'Select Window | ${ex}'")
+        except IndexError:
+            raise AssertionError("No window found")
+        except NoSuchWindowException:
+            raise AssertionError("Currently no focus window. where are you making a popup window?")
+        browser.switch_to_window(handles[-1])
 
-        try:
-            self._select_by_title(browser, criteria)
-            return
-        except ValueError: pass
-
-        raise ValueError("Unable to locate window with name or title '" + criteria + "'")
-
+    def _select_by_excludes(self, browser, excludes):
+        for handle in browser.get_window_handles():
+            if handle not in excludes:
+                browser.switch_to_window(handle)
+                return
+        raise ValueError("Unable to locate new window")
+    
     # Private
 
     def _parse_locator(self, locator):