+++ CHANGES.rst
@@ -3,6 +3,12 @@ Release Notes
 
 1.7 (unreleased)
 ----------------
+- Added keyword 'List Windows' to return a list of all window handles.
+- Enabled 'Select Window' to return window handle as well as accept it as locator, and
+  select new popup window by excluding previous window handles (the strict way) or
+  by special locator 'new' (the simplified but less strict way)
+  [divfor]
+
 - Added new keyword 'Wait Until Page Does Not Contain'.
   [deiga]